+++ pytorch_lightning/trainer/trainer.py
@@ -53,7 +53,7 @@ from pytorch_lightning.trainer.logging import TrainerLoggingMixin
 from pytorch_lightning.trainer.model_hooks import TrainerModelHooksMixin
 from pytorch_lightning.trainer.optimizers import TrainerOptimizersMixin
 from pytorch_lightning.trainer.properties import TrainerProperties
-from pytorch_lightning.trainer.states import TrainerState
+from pytorch_lightning.trainer.states import TrainerState, RunningStage
 from pytorch_lightning.trainer.training_loop import TrainLoop
 from pytorch_lightning.trainer.training_tricks import TrainerTrainingTricksMixin
 from pytorch_lightning.tuner.tuning import Tuner
@@ -921,3 +921,31 @@ class Trainer(
             Returns: List of all available plugins that are supported as string arguments.
         """
         return PluginConnector.available_plugins()
+
+    @property
+    def is_training(self) -> bool:
+        return self._running_stage == RunningStage.TRAINING
+
+    def set_training(self) -> None:
+        self._running_stage = RunningStage.TRAINING
+
+    @property
+    def is_testing(self) -> bool:
+        return self._running_stage == RunningStage.TESTING
+
+    def set_testing(self) -> None:
+        self._running_stage = RunningStage.TESTING
+
+    @property
+    def is_tuning(self) -> bool:
+        return self._running_stage == RunningStage.TUNING
+
+    def set_tuning(self) -> None:
+        self._running_stage = RunningStage.TUNING
+
+    @property
+    def is_evaluating(self) -> bool:
+        return self._running_stage == RunningStage.EVALUATING
+
+    def set_evaluating(self) -> None:
+        self._running_stage = RunningStage.EVALUATING
